expenditure

+++ Backend/src/modules/expenditure/services/DeleteExpenditure.service.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from 'tsyringe';
+import AppError from '../../../shared/errors/AppError';
 import IExpenditureRepository from '../interfaces/IExpenditureRepository';
 
 
@@ -9,7 +10,12 @@ export class DeleteExpenditureService {
         private repository: IExpenditureRepository
     ) { }
 
-    public async execute(id: number): Promise<void> {
+    public async execute(id: number, idCompany: number): Promise<void> {
+        const expediture = await this.repository.findByID(idCompany,id);
+
+        if(!expediture)
+            throw new AppError('Expenditure not found')
+
         await this.repository.delete(id);
     }
